# OnlineStore/OnlineStore/DTOs/Auth/RefreshTokenRequest.cs
using System.ComponentModel.DataAnnotations;

namespace OnlineStore.DTOs.Auth
{
    public class RefreshTokenRequest
    {
        [Required]
        public string AccessToken { get; set; } = string.Empty;

        [Required]
        public string RefreshToken { get; set; } = string.Empty;
    }
}